Move RuleMeta to historian/model package, iron out package dependencies
