+++ pkg/services/ngalert/state/historian/core.go
@@ -8,6 +8,7 @@ import (
 	"github.com/grafana/grafana/pkg/services/ngalert/eval"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 	"github.com/grafana/grafana/pkg/services/ngalert/state"
+	history_model "github.com/grafana/grafana/pkg/services/ngalert/state/historian/model"
 )
 
 func shouldRecord(transition state.StateTransition) bool {
@@ -36,7 +37,7 @@ type panelKey struct {
 }
 
 // panelKey attempts to get the key of the panel attached to the given rule. Returns nil if the rule is not attached to a panel.
-func parsePanelKey(rule state.RuleMeta, logger log.Logger) *panelKey {
+func parsePanelKey(rule history_model.RuleMeta, logger log.Logger) *panelKey {
 	if rule.DashboardUID != "" {
 		return &panelKey{
 			orgID:   rule.OrgID,
